L5_S1_HW/Tictactoe/requirements.txt
numpy==1.23.3
python-tictactoe==0.0.7
